requirement.txt
alembic==1.7.6
aniso8601==6.0.0
astroid==2.12.13
autopep8==1.7.0
Babel==2.9.0
click==8.0.3
dill==0.3.6
distlib==0.3.4
ecdsa        ==0.18.0
filelock==3.6.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.18.2
greenlet==1.1.2
importlib-metadata==4.12.0
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.1.1
python-jose=3.3.0
lazy-object-proxy==1.8.0
libxml2-python==2.9.12
lxml==4.9.1
Mako==1.1.6
MarkupSafe==2.1.1
mccabe==0.7.0
numpy==1.23.3
pandas==1.5.0
pip==22.2.2
platformdirs==2.5.1
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pycodestyle==2.9.1
pycryptodome==3.16.0
pylint==2.15.7
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
setuptools==60.5.0
six==1.12.0
SQLAlchemy==1.3.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
typed-ast==1.5.4
typing_extensions==4.4.0
virtualenv==20.13.4
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.1